pyproject.toml
[project]
name = "packagebot"
version = "0.1.0"
description = ""
authors = [
    {name = "packagebot"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "requests (>=2.32.5,<3.0.0)",
    "claude-agent-sdk (>=0.1.33,<0.2.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "temporalio (>=1.22.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.2"
black = "^26.1.0"
ruff = "^0.15.0"

